fix `reportInvalidCast` with variance (sort of, it still doesn't report invalid casts on non-overlapping generics due to https://github.com/DetachHead/basedpyright/issues/259)
